fix TripoSG dtype crash by casting the constant, not the activation
TripoSG Scribble died in the VAE decoder with

    expected mat1 and mat2 to have the same dtype, but got: float != c10::Half

FrequencyPositionalEmbedding registers `frequencies` with persistent=False,
so it is absent from the checkpoint. diffusers' torch_dtype loading sets
parameter dtypes from the weights and never re-.to()s the module, so that
buffer stayed fp32 while every weight around it was fp16. Multiplying
promoted the fp16 queries to fp32 for the rest of the decoder.

Most of this VAE is built from comfy.ops.manual_cast, which casts the weight
to the *activation* dtype per forward, so it absorbed the promotion silently
-- proj_query happily ran in fp32. The stock diffusers Attention did not: it
builds plain nn.Linear, which has no such adaptation and demands the input
match its fp16 weights. Hence a crash three layers downstream of the cause.

The fix follows the same rule as the ops: activations decide the dtype, and
anything that is not a weight is cast to them. _decode already pins queries
with .to(dtype=z.dtype); the embedder was the one place that undid it.

Checked that the embedding is now dtype-preserving for fp16/fp32/bf16 with
the buffer still fp32, and ran TripoSG_Scribble_Model.json end to end: the
node completes in 18s and writes real geometry (3.8 MB and 288 KB glb),
where before it raised at the first decode chunk.

Four sibling copies of this class have the identical fp32 buffer (PartCrafter,
craftsman, Hunyuan3D 2.1, Hunyuan3D V2). None of them fails today -- their
decoders are comfy ops throughout, so manual_cast absorbs the promotion the
whole way -- and casting there would change their numerics, so they are left
alone.

+++ nodes/Gen_3D_Modules/TripoSG/models/embeddings.py
@@ -85,7 +85,17 @@ class FrequencyPositionalEmbedding(nn.Module):
         """
 
         if self.num_freqs > 0:
-            embed = (x[..., None].contiguous() * self.frequencies).view(
+            # Cast the constant to the activation, never the other way round.
+            #
+            # `frequencies` is a persistent=False buffer, so it is absent from
+            # the checkpoint: diffusers' torch_dtype loading sets parameter
+            # dtypes from the weights and never re-.to()s the module, leaving
+            # this fp32 while everything around it is fp16. Multiplying then
+            # promoted fp16 queries to fp32, and the first layer that cannot
+            # adapt its weights -- stock diffusers Attention builds plain
+            # nn.Linear, unlike the ops.Linear the rest of the VAE uses --
+            # raised "expected mat1 and mat2 to have the same dtype".
+            embed = (x[..., None].contiguous() * self.frequencies.to(x.dtype)).view(
                 *x.shape[:-1], -1
             )
             if self.include_input:
